generalise manager
